restructure and variable activation function
