+++ src/mqt/predictor/ml/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from mqt.predictor.ml.graph_classifier import GNNClassifier, MultiGNNClassifier
+from mqt.predictor.ml.GNNclassifier import GNNClassifier, MultiGNNClassifier
 from mqt.predictor.ml.helper import qcompile
 from mqt.predictor.ml.Predictor import Predictor
